refactor: better file tree
