+++ src/globals.py
@@ -1,0 +1,52 @@
+import logging
+import os
+import socket
+from logging.handlers import RotatingFileHandler
+
+"""
+    server setup
+"""
+link = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
+link.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
+host = ''
+port = 12000
+link.bind((host, port))
+# list that will later contain the sockets
+clients = []
+
+"""
+    game data
+"""
+# determines whether the power of the character is used before
+# or after moving
+permanents, two, before, after = {'pink'}, {
+    'red', 'grey', 'blue'}, {'purple', 'brown'}, {'black', 'white'}
+# reunion of sets
+colors = before | permanents | after | two
+# ways between rooms
+passages = [{1, 4}, {0, 2}, {1, 3}, {2, 7}, {0, 5, 8},
+            {4, 6}, {5, 7}, {3, 6, 9}, {4, 9}, {7, 8}]
+# ways for the pink character
+pink_passages = [{1, 4}, {0, 2, 5, 7}, {1, 3, 6}, {2, 7}, {0, 5, 8, 9}, {4, 6, 1, 8}, {5, 7, 2, 9}, {3, 6, 9, 1},
+                 {4, 9, 5}, {7, 8, 4, 6}]
+
+"""
+    logging setup
+    you can set the appropriate importance level of the data
+    that are written to your logfiles.
+"""
+logger = logging.getLogger()
+logger.setLevel(logging.DEBUG)
+formatter = logging.Formatter(
+    "%(asctime)s :: %(levelname)s :: %(message)s", "%H:%M:%S")
+# logger to file
+if os.path.exists("./logs/game.log"):
+    os.remove("./logs/game.log")
+file_handler = RotatingFileHandler('./logs/game.log', 'a', 1000000, 1)
+file_handler.setLevel(logging.DEBUG)
+file_handler.setFormatter(formatter)
+logger.addHandler(file_handler)
+# logger to console
+stream_handler = logging.StreamHandler()
+stream_handler.setLevel(logging.DEBUG)
+logger.addHandler(stream_handler)
